// src/types.ts
export enum Buildings {
  URBAN_STRUCTURES = 'urbanStructures',
  SOLAR_PLANTS = 'solarPlants',
  GAS_PLANTS = 'gasPlants',
  FUSION_PLANTS = 'fusionPlants',
  ANTIMATTER_PLANTS = 'antimatterPlants',
  ORBITAL_PLANTS = 'orbitalPlants',
  RESEARCH_LABS = 'researchLabs',
  METAL_REFINERIES = 'metalRefineries',
  CRYSTAL_MINES = 'crystalMines',
  ROBOTIC_FACTORIES = 'roboticFactories',
  SHIPYARDS = 'shipyards',
  ORBITAL_SHIPYARDS = 'orbitalShipyards',
  SPACEPORTS = 'spaceports',
  COMMAND_CENTERS = 'commandCenters',
  NANITE_FACTORIES = 'naniteFactories',
  ANDROID_FACTORIES = 'androidFactories',
  ECONOMIC_CENTERS = 'economicCenters',
  TERRAFORM = 'terraform',
  MULTI_LEVEL_PLATFORMS = 'multiLevelPlatforms',
  ORBITAL_BASE = 'orbitalBase',
  JUMP_GATE = 'jumpGate',
  BIOSPHERE_MODIFICATION = 'biosphereModification',
  CAPITAL = 'capital',
}

export enum Technologies {
  ENERGY = 'energy',
  COMPUTER = 'computer',
  ARMOUR = 'armour',
  LASER = 'laser',
  MISSILES = 'missiles',
  STELLAR_DRIVE = 'stellarDrive',
  PLASMA = 'plasma',
  WARP_DRIVE = 'warpDrive',
  SHIELDING = 'shielding',
  ION = 'ion',
  PHOTON = 'photon',
  ARTIFICIAL_INTELLIGENCE = 'artificialIntelligence',
  DISRUPTOR = 'disruptor',
  CYBERNETICS = 'cybernetics',
  TACHYON_COMMUNICATIONS = 'tachyonCommunications',
  ANTI_GRAVITY = 'antiGravity',
}

export enum BaseStats {
  FERTILITY = 'fertility',
  AREA = 'area',
  GAS = 'gas',
  SOLAR = 'solar',
  METAL = 'metal',
  CRYSTAL = 'crystal',
  CONSTRUCTION = 'construction',
  PRODUCTION = 'production',
  ECONOMY = 'economy',
}

export enum Terrains {
  ARID = 'arid',
  ASTEROID = 'asteroid',
  CRATERS = 'craters',
  CRYSTALLINE = 'crystalline',
  EARTHLY = 'earthly',
  GAIA = 'gaia',
  GLACIAL = 'glacial',
  MAGMA = 'magma',
  METALLIC = 'metallic',
  OCEANIC = 'oceanic',
  RADIOACTIVE = 'radioactive',
  ROCKY = 'rocky',
  TOXIC = 'toxic',
  TUNDRA = 'tundra',
  VOLCANIC = 'volcanic',
}

export enum Ships {
  FIGHTER = 'fighter',
  BOMBER = 'bomber',
  HEEAVY_BOMBER = 'heavyBomber',
  ION_BOMBER = 'ionBomber',
  CORVETTE = 'corvette',
  RECYCLER = 'recycler',
  DESTROYER = 'destroyer',
  FRIGATE = 'frigate',
  ION_FRIGATE = 'ionFrigate',
  SCOUT_SHIP = 'scoutShip',
  OUTPOST_SHIP = 'outpostShip',
  CRUISER = 'cruiser',
  CARRIER = 'carrier',
  HEAVY_CRUISER = 'heavyCruiser',
  BATTLESHIP = 'battleShip',
  FLEET_CARRIER = 'fleetCarrier',
  DREADNOUGHT = 'dreadnought',
  TITAN = 'titan',
  LEVIATHAN = 'leviathan',
  DEATH_STAR = 'death_star',
}

export enum AstroTypes {
  MOON = 'moon',
  PLANET = 'planet',
}

export enum Defenses {
  BARRACKS = 'barracks',
  LASER_TURRETS = 'laserTurrets',
  MISSILE_TURRETS = 'missileTurrets',
  PLASMA_TURRETS = 'plasmaTurrets',
  ION_TURRETS = 'ionTurrets',
  PHOTON_TURRETS = 'photonTurrets',
  DISRUPTOR_TURRETS = 'disruptorTurrets',
  DEFLECTION_SHIELDS = 'deflectionShields',
  PLANETARY_SHIELDS = 'planetaryShield',
  PLANETARY_RING = 'planetaryRing',
}

export enum Weapons {
  LASER = 'laser',
  MISSILES = 'missiles',
  PLASMA = 'plasma',
  ION = 'ion',
  PHOTON = 'photon',
  DISRUPTOR = 'disruptor',
}